2017.07.26

+++ database/migrations/references/2017_06_27_163735_create_region.php
@@ -17,8 +17,7 @@ class CreateRegion extends Migration
             $table->unsignedSmallInteger('id')->nullable();
 			$table->primary('id');
 			$table->string('name', 50)->nullable();
-			//$table->boolean('is_parent')->nullable()->default(0);
-			$table->integer('parent_id')->default(0);
+			$table->unsignedSmallInteger('parent_id')->default(0);
 			$table->unsignedTinyInteger('city_id')->default(0);
 			$table->integer('weight')->default(1);
         });
